Extract camera utils namespace and place into sepearte src and header files.
