+++ src/camera_server_node.cpp
@@ -13,7 +13,6 @@
 #include <camera_driver/camera.h>
 #include <cv_bridge/cv_bridge.h>
 #include <image_transport/image_transport.h>
-#include <iostream>
 #include <ros/ros.h>
 
 #define encoding "bgr8"
